@@ -11,6 +11,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 
 namespace BusinessSample
 {
@@ -26,16 +27,17 @@ namespace BusinessSample
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
+            
             services.AddDbContextPool<BusinessDBContext>(options =>
             {
                 options.UseSqlServer(Configuration.GetConnectionString(config.BusinessDatabaseConnection));
             });
 
             //In Memory Connection for testing only
-            services.AddSingleton<IRestaurantData, InMemoryRestaurantData>();
+            //services.AddSingleton<IRestaurantData, InMemoryRestaurantData>();
 
             //SQL Connection
-            //services.AddScoped<IRestaurantData, SqlRestaurantData>();
+            services.AddScoped<IRestaurantData, SqlRestaurantData>();
 
             services.Configure<CookiePolicyOptions>(options =>
             {
@@ -43,11 +45,12 @@ namespace BusinessSample
                 options.CheckConsentNeeded = context => true;
                 options.MinimumSameSitePolicy = SameSiteMode.None;
             });
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
+            services.AddControllersWithViews();
+            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_3_0);
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
@@ -62,11 +65,21 @@ namespace BusinessSample
             app.Use(SayHelloMiddleWare);
 
             app.UseHttpsRedirection();
+            app.UseNodeModules(TimeSpan.FromSeconds(600));
             app.UseStaticFiles();
-            app.UseNodeModules(env);
-            app.UseCookiePolicy();
+            app.UseRouting();
 
-            app.UseMvc();
+            app.UseAuthorization();
+
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllerRoute
+                (   
+                    name: "default",
+                    pattern: "{controller=Home}/{action-Index}/{id?}"
+                );
+                //endpoints.MapRazorPages();
+            });
         }
 
         private RequestDelegate SayHelloMiddleWare(RequestDelegate next)
